Added Memory unit tests

+++ src/mem/mem_data.hpp
@@ -4,8 +4,9 @@
 
 class DataMemory {
 public:
-    int load(uint64_t addr, size_t bytes);
-    void store(uint64_t addr, size_t bytes, uint64_t val);
+  int64_t load(uint64_t addr, size_t bytes);
+  void store(uint64_t addr, size_t bytes, uint64_t val);
+
 private:
-    std::map<uint64_t, uint8_t> memory;
+  std::map<uint64_t, uint8_t> memory;
 };
